requirements.txt
# Facial Beauty Prediction (FBP) - geometric prior & training
# Python >= 3.9 recommended

numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
